test(plugin-state): split fresh-store read coverage into focused file
