+++ src/plugin-state/plugin-state-store.test.ts
@@ -1,5 +1,5 @@
 // Plugin state store tests cover per-plugin persisted state reads and writes.
-import { chmodSync, existsSync, mkdirSync, rmSync, statSync } from "node:fs";
+import { chmodSync, existsSync, rmSync, statSync } from "node:fs";
 import path from "node:path";
 import { DatabaseSync } from "node:sqlite";
 import { MAX_DATE_TIMESTAMP_MS } from "@openclaw/normalization-core/number-coercion";
@@ -922,49 +922,6 @@ describe("plugin state keyed store", () => {
         await expect(store.entries()).resolves.toEqual([]);
         expect(countPluginStateLiveEntries("discord", state.env)).toBe(0);
         expect(existsSync(databasePath)).toBe(false);
-      },
-    );
-  });
-
-  it("treats an existing state database without the lazy plugin-state table as empty", async () => {
-    await withOpenClawTestState(
-      { label: "plugin-state-read-only-table-missing", applyEnv: false },
-      async (state) => {
-        const databasePath = resolveOpenClawStateSqlitePath(state.env);
-        mkdirSync(path.dirname(databasePath), { recursive: true });
-        const database = new DatabaseSync(databasePath);
-        database.exec("CREATE TABLE unrelated_state (id INTEGER PRIMARY KEY)");
-        database.close();
-
-        const store = createPluginStateKeyedStore("discord", {
-          namespace: "read-only-table-missing",
-          maxEntries: 10,
-          env: state.env,
-        });
-
-        await expect(store.lookup("k")).resolves.toBeUndefined();
-        await expect(store.entries()).resolves.toEqual([]);
-        expect(
-          pluginStateEntriesInKeyRange({
-            pluginId: "discord",
-            namespace: "read-only-table-missing",
-            keyStartInclusive: "a",
-            keyEndExclusive: "z",
-            limit: 1,
-            env: state.env,
-          }),
-        ).toEqual([]);
-        expect(countPluginStateLiveEntries("discord", state.env)).toBe(0);
-
-        const verify = new DatabaseSync(databasePath, { readOnly: true });
-        expect(
-          verify
-            .prepare(
-              "SELECT name FROM sqlite_schema WHERE type = 'table' AND name = 'plugin_state_entries'",
-            )
-            .get(),
-        ).toBeUndefined();
-        verify.close();
       },
     );
   });
